Se puede generar envíos y consultar estado. Falta: - Creación txt provincias - Creación txt ordenes

+++ GrupoH - Proyecto/GrupoH - Proyecto/Validaciones.cs
@@ -62,7 +62,7 @@
             return stringIngresado;
         }
 
-        static public int ValidarIntIngresado(string mensaje, int min, int max)
+        static public int ValidarIntIngresado(string mensaje, int min, long max)
         {
 
             int intEntero;
